Fix parsing numbers in "plain" style
Number parsing was overly tolerant. Because NSScanner finishes parsing
integers once non-integer characters are found, -[NSScanner scanInt:]
was giving false positives.  Thus, resulting in hexadecimel or floating
point values to be parsed as integerse, with the remainder being
discarded (e.g. "43f4da" was parsed as an NSNumber with intValue 43).

The casting order was changed to prefer floating point numbers, and
ensures that the whole value is parsed, so as to parse digit-prefixed
strings properly.