Short-circuit the results if the tag specified is a String or Null tag.
There are two special tags, String and Null. When
a String tag (yml.org,2002:str) is specified then
we return the unmodified scalar value as NSString.

If the Null tag was specified, we don't have to
evaluate the value as the appropriate response is
always NSNull.

Include tests for explict String and Null casting.