Streamline casting process.
Before we try to cast a value from its implicit
casting to its explicit casting, lets make sure
that the source tag is not the same as the
destination tag.  If they are the same, then no
more processing is required -- just return the
value.

If they are not the same, then try to cast the
value.  If casting is not possible then return
the error value (NSNull).